# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Taxculate App OpenApi API Documentation",
        "description": "Taxculate App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "hdp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/countries": {
            "get": {
                "tags": [
                    "location"
                ],
                "summary": "Get all countries",
                "description": "Get all countries",
                "operationId": "all_countries",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/tax-transactions": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Post annual tax of an employee",
                "description": "Post annual tax of an employee",
                "operationId": "taxTransactions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaxTransactionCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TaxTransactionCreateFormRequest": {
                "title": "Tax Transaction Form Request Fields",
                "description": "Tax Transaction request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "taxClasses": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "basic_salary": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "housing": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "clothing": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "utility": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "lunch": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "education": {
                                    "type": "int",
                                    "example": "10000"
                                },
                                "vacation": {
                                    "type": "int",
                                    "example": "10000"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "taxDeductions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "nhf": {
                                    "type": "int",
                                    "example": true
                                },
                                "pension": {
                                    "type": "int",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "country_id": {
                        "title": "country_id",
                        "description": "country_id",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Taxculate Application",
            "description": "API Endpoints of Project"
        }
    ],
    "security": [
        []
    ]
}